+++ src/constants.js
@@ -15,5 +15,14 @@ module.exports = {
   // Used in cookie as well
   DEVICE_ID: 'amplitude_deviceId',
   OPT_OUT: 'amplitude_optOut',
-  USER_ID: 'amplitude_userId'
+  USER_ID: 'amplitude_userId',
+
+  COOKIE_TEST: 'amplitude_cookie_test',
+
+  // revenue keys
+  REVENUE_EVENT: 'revenue_amount',
+  REVENUE_PRODUCT_ID: '$productId',
+  REVENUE_QUANTITY: '$quantity',
+  REVENUE_PRICE: '$price',
+  REVENUE_REVENUE_TYPE: '$revenueType'
 };